Separated out get_system_tz
Separated out get_system_tz for easier reading of code and wrote a test for it.
