+++ inkycal/modules/inkycal_agenda.py
@@ -11,7 +11,9 @@ import calendar as cal
 
 from inkycal.modules.template import inkycal_module
 from inkycal.custom import *
+from inkycal.custom_functions.get_system_tz import get_system_tz
 from inkycal.modules.ical_parser import iCalendar
+
